chore: rename lib/gui/dot_env => lib/gui/DotEnv
