+++ src/apps/deskflow-gui/main.cpp
@@ -10,8 +10,8 @@
 #include "common/constants.h"
 #include "gui/Constants.h"
 #include "gui/Diagnostic.h"
+#include "gui/DotEnv.h"
 #include "gui/Logger.h"
-#include "gui/dotenv.h"
 #include "gui/messages.h"
 #include "gui/string_utils.h"
 #include "gui/style_utils.h"
