+++ src/test/integtests/gui/dotenv_tests.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: GPL-2.0-only WITH LicenseRef-OpenSSL-Exception
  */
 
-#include "gui/dotenv.h"
+#include "gui/DotEnv.h"
 
 #include <QFile>
 #include <QTextStream>
